move posture structs to enterprise module
